exercicio 40 refeiito código corrigido indice 0.3 renomeda classe  - treino
